[commit-21]: Changed TaskRepository and TaskTemplateRepository classes to interfaces. Added myBatis scripts. Added funtions in schema.sql

+++ src/main/java/com/zoola/taskmanager/domain/Task.java
@@ -17,5 +17,5 @@ public class Task {
     private int taskTemplateId;
     private Integer userId;
     private int assigneeId;
-    private int version;
+    private float version;
 }
